Update issue templates
Added templates for submitting issues to the repository.

The intent behind this is to create a uniform style for issue submission and ensure that as much detail and intent as possible is conveyed to other team members.

Submitting issues is one aspect of a well-thought-out and executed Git repo. Doing this will ensure that we meet the requirements detailed in the rubric regarding a clean repository.